+++ CMakeLists.txt
@@ -10,6 +10,8 @@ list(REMOVE_ITEM SOURCE ${REMOVE_CMAKE})
 
 add_executable(GPP ${SOURCE})
 
+#target_link_libraries(GPP "${CMAKE_SOURCE_DIR}/src/engine/audio/winmm.lib")
+
 set(CPACK_PROJECT_NAME ${PROJECT_NAME})
 set(CPACK_PROJECT_VERSION ${PROJECT_VERSION})
 include(CPack)
