+++ src/engine/display/display.cpp
@@ -1,0 +1,10 @@
+#include "display.h"
+
+#include <windows.h>
+
+/**
+ * @brief Clear screen
+ */
+void clrscr() {
+    system("cls");
+}
