+++ src/engine/display/display.h
@@ -1,0 +1,6 @@
+#ifndef DISPLAY_H
+#define DISPLAY_H
+
+void clrscr();
+
+#endif  // DISPLAY_H
