+++ src/engine/engine.h
@@ -1,13 +1,8 @@
 #ifndef ENGINE_H
 #define ENGINE_H
 
-// INCLUDES
-#include <cctype>
-#include <iostream>
-#include <string>
-
-void clrscr();
-std::string toLowerEach(std::string input);
-std::string stringPrompt(std::string message, bool lowercase);
+#include "display/display.h"
+#include "prompt/prompt.h"
+#include "text/text.h"
 
 #endif  // ENGINE_H
