+++ src/engine/text/text.h
@@ -1,0 +1,9 @@
+#ifndef TEXT_H
+#define TEXT_H
+
+#include <string>
+
+std::string toLowerEach(std::string input);
+std::string toUpperEach(std::string input);
+
+#endif  // TEXT_H
